// src/components/Select/interface.tsx
export interface SelectInputItem {
  /**
   * 实际值
   */
  value: string;
  /**
   * 展示值
   */
  label?: ReactNode;
}

export type SelectInputSize = 'small' | 'default' | 'large';

export interface SelectInputProps {
  /**
   * 当前的展示值
   */
  value?: ReactNode | SelectInputItem[];
  /**
   * 大小
   * @default 'default'
   */
  size?: SelectInputSize;
  /**
   * 是否禁用
   */
  disabled?: boolean;
  /**
   * 提示信息
   */
  placeholder?: string;
  /**
   * 是否允许清空
   * @default true
   */
  allowClear?: boolean;
  /**
   * 点击清空时候回调
   */
  onClear?: () => void;
  /**
   * 点击回调
   */
  onPress?: (event: GestureResponderEvent) => void;
  /**
   * 是否多个展示
   * @default false
   */
  multiple?: boolean;
  /**
   * 是否显示关闭 multiple为ture时候生效
   * @defalut false
   */
  closable?: boolean;
  /**
   * 点击关闭回调
   */
  onClose?: (item: SelectInputItem) => void;
  /**
   * 最外层容器样式
   */
  style?: StyleProp<ViewStyle>;
  /**
   * 自定义内容
   */
  content?: React.ReactElement;
  /**
   * 后缀插槽 可以自定义图标等
   */
  suffix?: ReactNode;
  /**
   * 自动化测试使用
   */
  testID?: string;
  /**
   * 校验错误状态 边框红色
   */
  status?: 'error';
}